Updated to .NET Core

+++ Utils/DMXrecorder/Common/DmxData.cs
@@ -19,7 +19,7 @@ namespace Animatroller.Common
 
         public byte[] Data { get; set; }
 
-        public ulong TimestampMS { get; set; }
+        public double TimestampMS { get; set; }
 
         public long Sequence { get; set; }
 
@@ -29,7 +29,7 @@ namespace Animatroller.Common
         {
         }
 
-        public static DmxData CreateNoChange(ulong millisecond, long sequence, int universe)
+        public static DmxData CreateNoChange(double millisecond, long sequence, int universe)
         {
             return new DmxData
             {
@@ -41,7 +41,7 @@ namespace Animatroller.Common
             };
         }
 
-        public static DmxData CreateFullFrame(ulong millisecond, long sequence, int universe, byte[] data)
+        public static DmxData CreateFullFrame(double millisecond, long sequence, int universe, byte[] data)
         {
             return new DmxData
             {
